Start cecie and command implementation

+++ internal/client/client.go
@@ -3,7 +3,7 @@ package client
 type ContentTarget struct {
 	TitleId string `yaml:"titleId"`
 	DirectoryName string `yaml:"directoryName"`
-	Blocks uint16 `yaml:"blocks"`
+	Blocks uint64 `yaml:"blocks"`
 }
 
 type Content struct {
